ssl_server/app/requirements.txt
asn1crypto==0.23.0
cffi==1.11.2
click==6.7
cryptography==2.1.1
Flask==0.12.2
idna==2.6
itsdangerous==0.24
Jinja2==2.9.6
MarkupSafe==1.0
pycparser==2.18
pyOpenSSL==17.3.0
six==1.11.0
Werkzeug==0.12.2
